DEV: Revision of NDArray (96% coverage)
* Add three derived classes of NDArray

The NDArray base and derived class are not intended to be used directly by the end user.
They serve as abstract classes that already implement the basic methods to define
spectral arrays and related attribute methods.

More specifically, derived classes such as NDDataset or Coord implement additional
methods and attributes.

What are the basic attributes of the NDArray class?
-----------------------------------------------------

The most used ones:
~~~~~~~~~~~~~~~~~~~
* `id`: A unique identifier of the NDArray object.
* `data`: An array of data contained in the object (default: None)
* `name`: A friendly name for the object (default: id)
* `dims`: The names of the dimensions.
* `quantity`: The quantity name of the data array (for example, `frequency`)
* `units`: The units of the data array (for example, `Hz`).

Useful additional attributes:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
* `meta`: A dictionary of any additional metadata.
* `timezone`: If not specified, the local timezone is assigned.

Read-only attributes:
~~~~~~~~~~~~~~~~~~~~~
* `dtype`: data type (see numpy definition of dtypes).

Most attributes are optional and automatically set by default.
Even data can be absent.  In this case, the NDArray is considered empty.

Derived classes
---------------
* NDComplexArray : This class has the same attributes as NDArray,
  but unlike NDArray, it accepts complex or hypercomplex data.

* NDMaskedComplexArray : This class has the same attributes as NDComplexArray,
  but has an additional attribute `mask`: NOMASK or a boolean array of the same form
  as the data array.

* NDLabelledArray : This class is a 1D NDArray, with an additional attribute `labels`.
  Labels can be defined in addition to the main data or replace it completely.

Backward compatibility
----------------------
For backward compatibility with version < 0.4, the `title` attribute is still
accessible, but has been deprecated as it is now replaced by the `quantity` attribute.

* Add tests for all methods and properties.